requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
stripe>=7.0
openai>=1.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.0
weasyprint>=60.0
django-ckeditor>=6.5
